+++ packages/client/src/observable/internal/list/ListQuery.ts
@@ -23,15 +23,10 @@ import type {
   PageResult,
   PropertyKeys,
 } from "@osdk/api";
-import groupBy from "object.groupby";
 import type { Observable, Subscription } from "rxjs";
 import invariant from "tiny-invariant";
-import { additionalContext, type Client } from "../../../Client.js";
+import { additionalContext } from "../../../Client.js";
 import type { InterfaceHolder } from "../../../object/convertWireToOsdkObjects/InterfaceHolder.js";
-import {
-  ObjectDefRef,
-  UnderlyingOsdkObject,
-} from "../../../object/convertWireToOsdkObjects/InternalSymbols.js";
 import type {
   ObjectHolder,
 } from "../../../object/convertWireToOsdkObjects/ObjectHolder.js";
@@ -534,156 +529,6 @@ export abstract class ListQuery extends BaseListQuery<
       });
     });
   }
-}
-
-export class ObjectListQuery extends ListQuery {
-  protected createObjectSet(store: Store): ObjectSet<ObjectTypeDefinition> {
-    return store.client({
-      type: "object",
-      apiName: this.apiName,
-    } as ObjectTypeDefinition)
-      .where(this.canonicalWhere);
-  }
-
-  async revalidateObjectType(apiName: string): Promise<void> {
-    if (this.apiName === apiName) {
-      await this.revalidate(/* force */ true);
-    }
-  }
-
-  protected postProcessFetchedData(
-    data: Osdk.Instance<any>[],
-  ): Promise<Osdk.Instance<any>[]> {
-    return Promise.resolve(data);
-  }
-
-  protected extractRelevantObjects(
-    changes: Changes,
-  ): ExtractRelevantObjectsResult {
-    return {
-      added: {
-        all: changes.addedObjects.get(this.cacheKey.otherKeys[API_NAME_IDX])
-          ?? [],
-        strictMatches: new Set(),
-        sortaMatches: new Set(),
-      },
-      modified: {
-        all: changes.modifiedObjects.get(this.cacheKey.otherKeys[API_NAME_IDX])
-          ?? [],
-        strictMatches: new Set(),
-        sortaMatches: new Set(),
-      },
-    };
-  }
-}
-
-export class InterfaceListQuery extends ListQuery {
-  protected createObjectSet(store: Store): ObjectSet<ObjectTypeDefinition> {
-    return store.client({
-      type: "interface" as const,
-      apiName: this.apiName,
-    } as ObjectOrInterfaceDefinition as ObjectTypeDefinition)
-      .where(this.canonicalWhere);
-  }
-
-  async revalidateObjectType(apiName: string): Promise<void> {
-    const objectMetadata = await this.store.client.fetchMetadata({
-      type: "object",
-      apiName,
-    });
-
-    if (this.apiName in objectMetadata.interfaceMap) {
-      await this.revalidate(/* force */ true);
-    }
-  }
-
-  protected async postProcessFetchedData(
-    data: Osdk.Instance<any>[],
-  ): Promise<Osdk.Instance<any>[]> {
-    return reloadDataAsFullObjects(this.store.client, data);
-  }
-
-  protected extractRelevantObjects(
-    changes: Changes,
-  ): ExtractRelevantObjectsResult {
-    const matchesApiName = ([, object]: [unknown, ObjectHolder]) => {
-      return this.apiName in object[ObjectDefRef].interfaceMap;
-    };
-
-    const added = Array.from(changes.addedObjects).filter(matchesApiName).map((
-      [, object],
-    ) => object.$as(this.apiName));
-
-    const modified = Array.from(changes.modifiedObjects).filter(matchesApiName)
-      .map((
-        [, object],
-      ) => object.$as(this.apiName));
-
-    return {
-      added: {
-        all: added,
-        strictMatches: new Set(),
-        sortaMatches: new Set(),
-      },
-      modified: {
-        all: modified,
-        strictMatches: new Set(),
-        sortaMatches: new Set(),
-      },
-    };
-  }
-}
-
-// Hopefully this can go away when we can just request the full object properties on first load
-async function reloadDataAsFullObjects(
-  client: Client,
-  data: Osdk.Instance<any>[],
-) {
-  const groups = groupBy(data, (x) => x.$objectType);
-  const objectTypeToPrimaryKeyToObject = Object.fromEntries(
-    await Promise.all(
-      Object.entries(groups).map<
-        Promise<
-          [
-            /** objectType **/ string,
-            Record<string | number, Osdk.Instance<ObjectTypeDefinition>>,
-          ]
-        >
-      >(async ([apiName, objects]) => {
-        // to keep InternalSimpleOsdkInstance simple, we make both the `ObjectDefRef` and
-        // the `InterfaceDefRef` optional but we know that the right one is on there
-        // thus we can `!`
-        const objectDef = (objects[0] as ObjectHolder)[UnderlyingOsdkObject][
-          ObjectDefRef
-        ]!;
-        const where: SimpleWhereClause = {
-          [objectDef.primaryKeyApiName]: {
-            $in: objects.map(x => x.$primaryKey),
-          },
-        };
-
-        const result = await client(
-          objectDef as ObjectTypeDefinition,
-        ).where(where).fetchPage();
-        return [
-          apiName,
-          Object.fromEntries(result.data.map(
-            x => [x.$primaryKey, x],
-          )),
-        ];
-      }),
-    ),
-  );
-
-  data = data.map((obj) => {
-    invariant(
-      objectTypeToPrimaryKeyToObject[obj.$objectType][obj.$primaryKey],
-      `Could not find object ${obj.$objectType} ${obj.$primaryKey}`,
-    );
-    return objectTypeToPrimaryKeyToObject[obj.$objectType][obj.$primaryKey];
-  });
-
-  return data;
 }
 
 export function isListCacheKey(
